Use Nosetests for the Django unit / developer tests
At this time, we have to install the test requirements onto the base
Docker image. This is suboptimal as it needlessly inflates the
container size, and is not security best practice either. However,
the only alternative appears to be to leave these packages out of the
Docker build, and instead install them each time the Fig container is
brought up for dev tests. This represents a large time overhead for such
a common action. There is a third alternative which appears currently
to be unavailable: That is, to add the test packages to the Fig
configuration. There appears to be no way to do this at the current
version of Fig. I have raised an issue on github:
https://github.com/docker/fig/issues/600

Note also that the "USE_NOSETESTS" environment var is set in the Fig
container; this prompts Django to include django_nose in its apps and
to change the test runner. Although not required by the dev server
(which is the primary command of the Fig container), there are required
by the dev tests, and so this var is set at the Fig container level
for convenience of test runs.