Ensure an empty Postgres container for each dev browser test
This is pretty hacky but seemingly the only way to get this done in
Fig. The standard Postgres Docker image persists its data by
sharing a Volume with the Docker host. For these tests, we don't
want persistence between invocations of the image as a container.
However there seems to be no way to override this with Fig
currently, so here is a hacky bash way to kill of the container
after each test.